requirements.txt
colorama==0.3.7
envoy==0.0.3
python-dateutil==2.4.1
sinaweibopy